pyproject.toml
[tool.poetry]
name = "vianu"
version = "0.0.1"
description = "Vianu enables developers to access tools applicable to life sciences and healthcare."
authors = [
    "Nicolas Perez Gonzalez <hello@vianu.org>",
    "Alexander Horst <hello@vianu.org>"
]
repository = "https://github.com/smc40/vianu"
license = "MIT"
keywords = ["healthcare", "government", "opensource", "data science"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
gradio = "^4.0"
build = "^1.2.2.post1"
twine = "^5.1.1"
abydos = "^0.5.0"
cologne-phonetics = "^2.0.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
llama-cpp-python = "^0.3.2"

[tool.poetry.scripts]
nightcrawler_app = "vianu.tools.nightcrawler.launch_demo_app:main"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools]
exclude-patterns = [".env"]